% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality_parsers.R
\name{filter_SNVs}
\alias{filter_SNVs}
\title{Filters SNVdata in tumordata object}
\usage{
filter_SNVs(td, filter_min_DP = 0, filter_min_alt_reads = 0)
}
\arguments{
\item{td}{(required) tumordata object}

\item{filter_min_DP}{minimal sequencing depth of mutation required in at least one sample}

\item{filter_min_alt_reads}{minimal alt_reads of mutation required in at least one sample}
}
\value{
tumordata object
}
\description{
Filters SNVdata in tumordata object
}
\examples{
data(td)
filter(td, filter_min_DP = 10, filter_min_alt_reads = 0)
}
