% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality_parsers.R
\name{prepare_pycloneVI_input}
\alias{prepare_pycloneVI_input}
\title{Pyclone_VI input parser}
\usage{
prepare_pycloneVI_input(vcf_file, cnv_files, sample_ids, sex, genome_build)
}
\arguments{
\item{vcf_file}{Path to VCF file}

\item{cnv_files}{Path to CNV files}

\item{sample_ids}{Sample IDs}

\item{sex}{"male"/"female"}

\item{genome_build}{eg. "hg38"}
}
\value{
parsed tibble
}
\description{
Pyclone_VI input parser
}
